pyproject.toml
[project]
name = "health_orchestrator"
version = "0.1.0"
description = "Governance-aware healthcare orchestration engine"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "pydantic",
    "langgraph>=1.0.9",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
]

[tool.uv]
